+++ src/Validation/Validation.ts
@@ -20,7 +20,8 @@ import {
     IValidationReplaceDto,
     IValidationIsjwtDto,
     IRequiredValidationDto,
-    IValidationIsStrongPasswordDto
+    IValidationIsStrongPasswordDto,
+    IValidationIsEmailDto
 } from "./dtos";
 
 import { get,isEqual } from "lodash";
@@ -1031,6 +1032,48 @@ export class Validation {
 
         if (params.pattern != null && !(params.pattern instanceof RegExp)) {
             throw new Error(`(isStrongPassword) validation pattern must be a regex`)
+        }
+
+        return (field : string) => {
+
+            const toMatch = Util.returnBasedOnCheckIn(checkIn,field);
+
+            message = Util.replaceMessageWithField(field, message);
+
+            return toMatch
+                .custom((value) => {
+                    if (params.pattern !== undefined) {
+
+                        return params.pattern.test(value)
+                            ? Promise.resolve()
+                            : Promise.reject(message)
+                    }
+                })
+                .withMessage((value : unknown) => message.replace(/(:value)|(:data)/ig,`${value}`));
+        }
+
+    }
+
+    /**
+     * To check if the string is a valid email address
+     *
+     * @param {IValidationIsEmailDto} [validation_options]
+     */
+    static isEmail(validation_options: IValidationIsEmailDto = {}) : Function {
+
+        const {
+            checkIn = "any",
+            params = {
+                pattern: /^[^\s@]+@[^\s@]+\.[^\s@]+$/
+            }
+        } = validation_options;
+
+        let {
+            message = `The :attribute is not a valid email address`
+        } = validation_options;
+
+        if (params.pattern != null && !(params.pattern instanceof RegExp)) {
+            throw new Error(`(isEmail) validation pattern must be a regex`)
         }
 
         return (field : string) => {