+++ src/Validation/dtos/index.ts
@@ -18,3 +18,4 @@ export { IValidationTrimDto } from "./validation.trim.dto";
 export { IValidationReplaceDto } from "./validation.replace.dto";
 export { IValidationIsjwtDto } from "./validation.isjwt.dto";
 export { IValidationIsStrongPasswordDto } from "./validation.isstrongpassword.dto"
+export { IValidationIsEmailDto } from "./validation.isemail.dto"