+++ src/Validation/dtos/validation.isemail.dto.ts
@@ -1,0 +1,7 @@
+import { ValidationGeneralDto } from "./validation.general.dto";
+
+export interface IValidationIsEmailDto extends Omit<ValidationGeneralDto, 'customFunction' | 'params'> {
+    params?: {
+        pattern?: RegExp
+    }
+}